+++ prov/rxm/src/rxm.h
@@ -49,7 +49,7 @@
 #include <fi_enosys.h>
 #include <fi_util.h>
 #include <fi_list.h>
-#include <fi_proto.h>
+#include <ofi_proto.h>
 #include <ofi_iov.h>
 
 #ifndef _RXM_H_
